+++ app/controllers/api/v2/pages_controller.rb
@@ -55,6 +55,10 @@ class Api::V2::PagesController < ApiController
     authorize! :author, space,
       message: "You need to be an editor of this space to edit pages."
 
+    unless Lux::Lock.for(page).acquire!(holder: current_user) == Lux::LOCK_OK
+      return head 409
+    end
+
     xparams = params.require(:page).permit(:title, :content, :browsable, :folder_id, :encrypted, :digest)
 
     if new_folder_id = xparams[:folder_id]
@@ -91,12 +95,18 @@ class Api::V2::PagesController < ApiController
   def destroy
     page = Page.find(params[:page_id])
 
+    unless Lux::Lock.for(page).acquire!(holder: current_user) == Lux::LOCK_OK
+      return head 409
+    end
+
     authorize! :delete, page,
       message: "You can not remove pages authored by someone else."
 
     unless page.destroy
       halt! 500, page.errors
     end
+
+    Lux::Lock.for(page).release!(holder: current_user)
 
     head 204
   end
